backend/api/requirements.txt
requests>=2.31.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
pyyaml>=6.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
argparse
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
transformers>=4.30.0
timm>=0.9.0
huggingface_hub>=0.16.0python-dotenv>=0.19.0
